requirements.txt
Django>=2.0,<3.0
djangorestframework
markdown
django-filter
django-csvimport
django-widget-tweaks
pillow
pytest
pytest-factoryboy
pytest-django